+++ Gruntfile.js
@@ -35,14 +35,14 @@ module.exports = function (grunt) {
           sourceMap: true,
           sourceMapName: 'dist/angular-cache.min.map',
           banner: '/**\n' +
-            '* @author Jason Dobry <jason.dobry@gmail.com>\n' +
-            '* @file angular-cache.min.js\n' +
-            '* @version <%= pkg.version %> - Homepage <https://github.com/jmdobry/angular-cache>\n' +
-            '* @copyright (c) 2013-2014 Jason Dobry <http://www.pseudobry.com>\n' +
-            '* @license MIT <https://github.com/jmdobry/angular-cache/blob/master/LICENSE>\n' +
-            '*\n' +
-            '* @overview angular-cache is a very useful replacement for Angular\'s $cacheFactory.\n' +
-            '*/\n'
+          '* @author Jason Dobry <jason.dobry@gmail.com>\n' +
+          '* @file angular-cache.min.js\n' +
+          '* @version <%= pkg.version %> - Homepage <https://github.com/jmdobry/angular-cache>\n' +
+          '* @copyright (c) 2013-2014 Jason Dobry <http://www.pseudobry.com>\n' +
+          '* @license MIT <https://github.com/jmdobry/angular-cache/blob/master/LICENSE>\n' +
+          '*\n' +
+          '* @overview angular-cache is a very useful replacement for Angular\'s $cacheFactory.\n' +
+          '*/\n'
         },
         files: {
           'dist/angular-cache.min.js': ['dist/angular-cache.js']
@@ -158,19 +158,38 @@ module.exports = function (grunt) {
     grunt.file.write(filePath, file);
   });
 
+  grunt.registerTask('banner', function () {
+    var file = grunt.file.read('dist/angular-cache.js');
+
+    var banner = '/**\n' +
+      '* @author Jason Dobry <jason.dobry@gmail.com>\n' +
+      '* @file angular-cache.js\n' +
+      '* @version ' + pkg.version + ' - Homepage <https://github.com/jmdobry/angular-cache>\n' +
+      '* @copyright (c) 2013-2014 Jason Dobry <http://www.pseudobry.com>\n' +
+      '* @license MIT <https://github.com/jmdobry/angular-cache/blob/master/LICENSE>\n' +
+      '*\n' +
+      '* @overview angular-cache is a very useful replacement for Angular\'s $cacheFactory.\n' +
+      '*/\n';
+
+    file = banner + file;
+
+    grunt.file.write('dist/angular-cache.js', file);
+  });
+
   grunt.registerTask('test', ['karma:1.2.25']);
   grunt.registerTask('build', [
     'clean',
     'jshint',
     'browserify',
     'version:dist/angular-cache.js',
+    'banner',
     'uglify'
   ]);
   grunt.registerTask('default', ['build']);
 
   // Used by the CLI build servers
   grunt.registerTask('test-ci', [
-		'karma:1.1.5',
+    'karma:1.1.5',
     'karma:1.2.16',
     'karma:1.2.25',
     'karma:min',